// ClientApp/src/app/models/docente.ts
export class Docente{
    Id: number;
    Tipo_Documento: string;
    Identificacion: string;
    Primer_Nombre: string;
    Segundo_Nombre: string;
    Primer_Apellido: string;
    Segundo_Apellido: string;
    Fecha_Nacimiento: string;
    Edad: number;
    Genero: string;
    Email: string;
    Telefono: number;
    Cargo: string;
    Fecha_Vinculacion: string;
    EstadoSys: string;
    Tipo_Docente: string;
}